Move runtime into modules
This moves the traceur runtime into modules inside runtime/modules.
This is in preparation for allowing importing the runtime as
modules in the future.
