+++ src/runtime/exportStar.js
@@ -12,24 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-const {defineProperty, getOwnPropertyNames} = Object;
-
-function exportStar(object) {
-  for (let i = 1; i < arguments.length; i++) {
-    let mod = arguments[i];
-    let names = getOwnPropertyNames(mod);
-    for (let j = 0; j < names.length; j++) {
-      let name = names[j];
-      if (name === '__esModule' || name === 'default') {
-        continue;
-      }
-      defineProperty(object, name, {
-        get: () => mod[name],
-        enumerable: true
-      });
-    }
-  }
-  return object;
-}
-
+import exportStar from './modules/exportStar.js';
 $traceurRuntime.exportStar = exportStar;
