+++ src/runtime/modules/iteratorToArray.js
@@ -12,10 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import asyncWrap from './modules/asyncWrap.js';
-import initGeneratorFunction from './modules/initGeneratorFunction.js';
-import createGeneratorInstance from './modules/createGeneratorInstance.js';
-
-$traceurRuntime.asyncWrap = asyncWrap;
-$traceurRuntime.initGeneratorFunction = initGeneratorFunction;
-$traceurRuntime.createGeneratorInstance = createGeneratorInstance;
+export default function iteratorToArray(iter) {
+  var rv = [];
+  var i = 0;
+  var tmp;
+  while (!(tmp = iter.next()).done) {
+    rv[i++] = tmp.value;
+  }
+  return rv;
+}
